Update server
The server is now multi-thread, using thread pool to handle client
connections. The drawer page let the client to draw digits on the
canvas, and by pressing the predict button, it sends the pixel data
to server, server then use the model (neuralib.h) to predict the
digit. Finally, server sends back the result to client.